requirements.txt
sjvisualizer>=0.0.15
yfinance>=0.2.0
pandas>=2.0.0
openpyxl>=3.1.0
PyYAML>=6.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
requests>=2.31.0
